// src/types/PokemonType.ts
export type PokemonType = {
  name: string;
  nameJa: string;
  flavorText: string;
  general: string;
  height: string;
  weight: string;
  img: {
    default: string;
    defaultBack: string;
    shiny: string;
    shinyBack: string;
    official: string;
  };
  animate: {
    default: string;
    defaultBack: string;
    shiny: string;
    shinyBack: string;
  };
};